правки по ревью

+++ src/test/java/unauthorizedUser/NewUserRegistrationTest.java
@@ -2,64 +2,68 @@ package unauthorizedUser;
 
 
 import org.apache.commons.lang3.RandomStringUtils;
+import org.junit.After;
+import org.junit.Before;
 import org.junit.Rule;
 import org.junit.Test;
-import org.openqa.selenium.By;
-import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.html5.LocalStorage;
 import org.openqa.selenium.html5.WebStorage;
-import org.openqa.selenium.support.ui.ExpectedConditions;
-import org.openqa.selenium.support.ui.WebDriverWait;
 import stellarBurger.api.UserClient;
 import stellarBurger.pages.LoginPage;
 import stellarBurger.pages.MainPageBurgers;
 import stellarBurger.pages.RegistrationFormPage;
-
-import java.time.Duration;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotNull;
 
 public class NewUserRegistrationTest {
     private final UserClient client = new UserClient();
+    private final String name = RandomStringUtils.randomAlphabetic(10);
+    private final String email = RandomStringUtils.randomAlphanumeric(10) + "@yandex.ru";
+    private final String correctPassword = RandomStringUtils.randomAlphanumeric(10);
+    private final String incorrectPassword = RandomStringUtils.randomAlphanumeric(5);
     @Rule
     public DriverRule driverRule = new DriverRule();
+    private String accessToken;
+    private RegistrationFormPage registrationPage;
+
+    @Before
+    public void setUp() {
+        registrationPage = new RegistrationFormPage(driverRule.getDriver());
+        registrationPage.open()
+                .waitForLoadRegistrationForm();
+    }
+
+    @After
+    public void deleteUser() {
+        if (accessToken != null) {
+            client.delete(accessToken);
+        }
+    }
 
     @Test
     public void registerNewUser() {
-        WebDriver driver = driverRule.getDriver();
-        driver.get("https://stellarburgers.nomoreparties.site/register");
-        String name = RandomStringUtils.randomAlphabetic(10);
-        String email = RandomStringUtils.randomAlphanumeric(10) + "@yandex.ru";
-        String password = RandomStringUtils.randomAlphanumeric(10);
-        RegistrationFormPage objRegistrationPage = new RegistrationFormPage(driver);
-        objRegistrationPage.fillRegistrationForm(name, email, password);
-        objRegistrationPage.clickSignUpButton();
-        LoginPage objLogin = new LoginPage(driver);
-        objLogin.waitForLoadLoginPage();
-        objLogin.fillLoginForm(email, password);
-        objLogin.clickLogInButton();
-        MainPageBurgers objMainPage = new MainPageBurgers(driver);
-        objMainPage.waitForLoadMainPage();
-        LocalStorage localStorage = ((WebStorage) driver).getLocalStorage();
-        String accessToken = localStorage.getItem("accessToken");
+        registrationPage.fillRegistrationForm(name, email, correctPassword);
+
+        LoginPage loginPage = registrationPage.clickSignUpButton();
+        loginPage.waitForLoadLoginPage()
+                .fillLoginForm(email, correctPassword);
+
+        MainPageBurgers mainPage = loginPage.clickLogInButton();
+        mainPage.waitForLoadMainPage();
+
+        LocalStorage localStorage = ((WebStorage) driverRule.getDriver()).getLocalStorage();
+        accessToken = localStorage.getItem("accessToken");
         assertNotNull(accessToken);
-        client.delete(accessToken);
     }
 
     @Test
     public void registrationWithPasswordLessThenSixSymbols() {
-        WebDriver driver = driverRule.getDriver();
-        driver.get("https://stellarburgers.nomoreparties.site/register");
-        String password = RandomStringUtils.randomAlphanumeric(5);
-        var errorMessage = By.cssSelector(".input__error");
-        RegistrationFormPage objRegistrationPage = new RegistrationFormPage(driver);
-        objRegistrationPage.setPassword(password);
-        objRegistrationPage.clickSignUpButton();
-        new WebDriverWait(driver, Duration.ofSeconds(10))
-                .until(ExpectedConditions.visibilityOfElementLocated(errorMessage));
+        registrationPage.setPassword(incorrectPassword);
+        registrationPage.clickSignUpButton();
+        registrationPage.waitForErrorMessage();
         String expectedErrorMessage = "Некорректный пароль";
-        String actualErrorMessage = driver.findElement(errorMessage).getText();
+        String actualErrorMessage = registrationPage.getPasswordErrorText();
         assertEquals(expectedErrorMessage, actualErrorMessage);
     }
 }
